Update glusterfs server on RHEL 8
Because install glusterfs-server fail on RHEL 8. So change to
use glusterfs server on download.libvirt.redhat.com for RHEL 8.

Reviewed-on: https://code.engineering.redhat.com/gerrit/166513
Tested-by: Libvirt Jenkins <libvirt-jenkins@redhat.com>

+++ repos/installation/install_common.py
@@ -36,8 +36,13 @@ def setup_storage(params, mountpath, logger):
         if storage == "gluster":
             if not os.path.isdir(brickpath):
                 os.mkdir(brickpath, 0o755)
-            utils.setup_gluster("test-api-gluster", utils.get_local_hostname(), brickpath, logger)
-            utils.mount_gluster("test-api-gluster", utils.get_local_hostname(), mountpath, logger)
+            if utils.isRelease("8", logger):
+                gluster_server = "download.libvirt.redhat.com"
+                utils.set_fusefs(logger)
+            else:
+                gluster_server = utils.get_local_hostname()
+                utils.setup_gluster("test-api-gluster", gluster_server, brickpath, logger)
+            utils.mount_gluster("test-api-gluster", gluster_server, mountpath, logger)
             diskpath = mountpath + "/" + imagename
             remove_all(diskpath, logger)
             create_image(diskpath, seeksize, imageformat, logger)
@@ -58,9 +63,10 @@ def cleanup_storage(params, mountpath, logger):
     sourcepath = params.get('sourcepath')
     if storage == "gluster":
         utils.umount_gluster(mountpath, logger)
-        utils.cleanup_gluster("test-api-gluster", logger)
-        if os.path.isdir(brickpath):
-            shutil.rmtree(brickpath)
+        if not utils.isRelease("8", logger):
+            utils.cleanup_gluster("test-api-gluster", logger)
+            if os.path.isdir(brickpath):
+                shutil.rmtree(brickpath)
     elif storage == "nfs":
         utils.cleanup_nfs(mountpath, logger)
     elif storage == "iscsi":
